fix issue #17: hyprpicker stops working after a couple of uses
Hyprpicker stopped working after running several times when started
with option '-a'. The reason was that finish() has not been called after
wl-copy replaced the currently running process due to the call of
execlp() (execlp() never returns in that case).

To fix this issue finish now is called before the call of execlp()
when hyprpicker is running with option '-a'